Move bughunt report tests to modules
