+++ crates/runner/src/bughunt/report.rs
@@ -375,3 +375,228 @@ impl BughuntReport {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn metadata(mode: &str) -> RunMetadata {
+        RunMetadata {
+            mode: mode.to_string(),
+            model: "qwen3.5-0.8b".to_string(),
+            backend: "metal".to_string(),
+            device: 0,
+            arch: "apple-m4".to_string(),
+            model_dir: "/tmp/model".to_string(),
+            oracle_device: "cpu".to_string(),
+            commit_ish: Some("abc123".to_string()),
+        }
+    }
+
+    #[test]
+    fn report_serialization_includes_localization_fields() {
+        let report = BughuntReport {
+            mode: "localize".to_string(),
+            metadata: metadata("localize"),
+            gate: None,
+            decode_gate: None,
+            localize: Some(LocalizeRunSection {
+                pass: false,
+                gate_prompt: PromptGateReport {
+                    name: "code_prompt".to_string(),
+                    notes: Some("code".to_string()),
+                    pass: false,
+                    thresholds: PromptThresholds {
+                        prefill_logit_max_abs: 0.1,
+                        layer_hidden_max_abs: 0.1,
+                        restart_tail_logit_max_abs: 0.1,
+                    },
+                    prefill_logit_reference: "oracle_final_hidden_recomputed".to_string(),
+                    prefill_logit_max_abs: 0.2,
+                    prefill_logit_mean_abs: 0.02,
+                    prefill_logit_mse: 0.01,
+                    raw_oracle_prefill_logit_max_abs: 0.25,
+                    gpu_reference_logit_max_abs: 0.19,
+                    native_vs_gpu_reference_logit_max_abs: 0.03,
+                    worst_checked_position: 15,
+                    worst_layer: 18,
+                    worst_layer_kind: "linear".to_string(),
+                    worst_layer_delta: 0.12,
+                    checked_positions: Vec::new(),
+                    timings: vec![PhaseTimingReport {
+                        phase: "native_prefill".to_string(),
+                        elapsed_ms: 12.5,
+                    }],
+                },
+                localization: LocalizationSummary {
+                    prompt_name: "code_prompt".to_string(),
+                    initial_suspicious_position: 15,
+                    initial_suspicious_layer: 18,
+                    initial_suspicious_layer_kind: "linear".to_string(),
+                    per_layer_hidden_sweep: Vec::new(),
+                    restart_layer_sweep: Vec::new(),
+                    first_suspicious_restart_layer: Some(18),
+                    restart_position_scan: Vec::new(),
+                    worst_sampled_position: Some(15),
+                    chosen_traced_layer: Some(18),
+                    chosen_traced_layer_kind: Some("linear".to_string()),
+                    traced_metrics: None,
+                },
+            }),
+            dump: None,
+            bench: None,
+        };
+        let value = serde_json::to_value(&report).unwrap();
+        assert_eq!(value["mode"], "localize");
+        assert_eq!(
+            value["localize"]["localization"]["first_suspicious_restart_layer"],
+            18
+        );
+    }
+
+    #[test]
+    fn bench_report_serialization_includes_decode_and_profile_fields() {
+        let report = BughuntReport {
+            mode: "bench".to_string(),
+            metadata: metadata("bench"),
+            gate: None,
+            decode_gate: None,
+            localize: None,
+            dump: None,
+            bench: Some(BenchRunSection {
+                pass: true,
+                prompt_results: vec![BenchPromptReport {
+                    name: "hello_world".to_string(),
+                    notes: Some("smoke".to_string()),
+                    prompt_len: 2,
+                    warmup_iterations: 0,
+                    iterations: 1,
+                    decode_tokens: 1,
+                    native_prefill_ms: vec![3.0],
+                    min_native_prefill_ms: 3.0,
+                    max_native_prefill_ms: 3.0,
+                    mean_native_prefill_ms: 3.0,
+                    greedy_prefill_ms: vec![1.0],
+                    min_greedy_prefill_ms: 1.0,
+                    max_greedy_prefill_ms: 1.0,
+                    mean_greedy_prefill_ms: 1.0,
+                    replay_decode_ms: vec![4.0],
+                    min_replay_decode_ms: Some(4.0),
+                    max_replay_decode_ms: Some(4.0),
+                    mean_replay_decode_ms: Some(4.0),
+                    mean_replay_decode_ms_per_token: Some(4.0),
+                    component_decode_ms: Some(vec![2.0]),
+                    min_component_decode_ms: Some(2.0),
+                    max_component_decode_ms: Some(2.0),
+                    mean_component_decode_ms: Some(2.0),
+                    mean_component_decode_ms_per_token: Some(2.0),
+                    prefill_profile: Some(MetalProfileReport {
+                        total_calls: 2,
+                        native_calls: 1,
+                        host_calls: 1,
+                        total_ms: 1.5,
+                        native_ms: 1.0,
+                        host_ms: 0.5,
+                        entries: vec![MetalProfileOpReport {
+                            op: "matmul_rhs_transposed".to_string(),
+                            path: "native".to_string(),
+                            calls: 1,
+                            total_ms: 1.0,
+                            mean_ms: 1.0,
+                            max_ms: 1.0,
+                        }],
+                    }),
+                    greedy_prefill_profile: None,
+                    replay_decode_profile: None,
+                    component_decode_profile: None,
+                    prefill_hal_profile: Some(HalProfileReport {
+                        total_calls: 3,
+                        total_ms: 2.0,
+                        alloc_calls: 1,
+                        alloc_bytes: 1024,
+                        free_calls: 1,
+                        h2d_bytes: 0,
+                        d2h_bytes: 2048,
+                        d2d_bytes: 0,
+                        memset_bytes: 1024,
+                        sync_calls: 1,
+                        entries: vec![HalProfileOpReport {
+                            op: "alloc".to_string(),
+                            calls: 1,
+                            total_ms: 1.0,
+                            mean_ms: 1.0,
+                            max_ms: 1.0,
+                            total_bytes: 1024,
+                        }],
+                    }),
+                    greedy_prefill_hal_profile: None,
+                    replay_decode_hal_profile: None,
+                    component_decode_hal_profile: None,
+                }],
+            }),
+        };
+        let value = serde_json::to_value(&report).unwrap();
+        let prompt = &value["bench"]["prompt_results"][0];
+        assert_eq!(prompt["decode_tokens"], 1);
+        assert_eq!(prompt["mean_greedy_prefill_ms"], 1.0);
+        assert_eq!(prompt["mean_replay_decode_ms_per_token"], 4.0);
+        assert_eq!(prompt["mean_component_decode_ms_per_token"], 2.0);
+        assert_eq!(prompt["prefill_profile"]["native_calls"], 1);
+        assert_eq!(
+            prompt["prefill_profile"]["entries"][0]["op"],
+            "matmul_rhs_transposed"
+        );
+        assert_eq!(prompt["prefill_hal_profile"]["alloc_calls"], 1);
+        assert_eq!(prompt["prefill_hal_profile"]["d2h_bytes"], 2048);
+    }
+
+    #[test]
+    fn decode_gate_report_serialization_includes_token_comparisons() {
+        let report = BughuntReport {
+            mode: "decode_gate".to_string(),
+            metadata: metadata("decode_gate"),
+            gate: None,
+            decode_gate: Some(DecodeGateRunSection {
+                pass: false,
+                prompt_results: vec![DecodeGatePromptReport {
+                    name: "code_prompt".to_string(),
+                    notes: Some("code".to_string()),
+                    pass: false,
+                    prompt_len: 31,
+                    decode_tokens: 2,
+                    oracle_tokens: vec![271, 16],
+                    replay_tokens: vec![271, 17],
+                    component_tokens: Some(vec![271, 16]),
+                    max_replay_logit_abs: 0.2,
+                    mean_replay_logit_abs: 0.05,
+                    first_mismatch_step: Some(1),
+                    steps: vec![DecodeStepGateReport {
+                        step: 1,
+                        oracle_token: 16,
+                        replay_token: 17,
+                        component_token: Some(16),
+                        replay_logit_reference: "oracle_decode_logits".to_string(),
+                        replay_logit_max_abs: 0.2,
+                        replay_logit_mean_abs: 0.05,
+                        token_match_replay: false,
+                        token_match_component: true,
+                    }],
+                    timings: vec![PhaseTimingReport {
+                        phase: "oracle".to_string(),
+                        elapsed_ms: 10.0,
+                    }],
+                }],
+            }),
+            localize: None,
+            dump: None,
+            bench: None,
+        };
+        let value = serde_json::to_value(&report).unwrap();
+        let prompt = &value["decode_gate"]["prompt_results"][0];
+        assert_eq!(value["mode"], "decode_gate");
+        assert_eq!(prompt["first_mismatch_step"], 1);
+        assert_eq!(prompt["oracle_tokens"][1], 16);
+        assert_eq!(prompt["replay_tokens"][1], 17);
+        assert_eq!(prompt["component_tokens"][1], 16);
+    }
+}
